Update and refactor remaining make request about case to court tests
