--- e2e/fixtures/citizen/caseView/makeRequestToCourtAboutCase/applicant/requestToOrderWitnessContent.ts
@@ -1,9 +1,0 @@
-export enum RequestToOrderWitnessContent1 {
-  GovukHeadingXl = "Make a request to order a witness to attend court",
-  GovukBody1 = "You will need to pay a fee of £0.00 for your request to ask the court to order a witness to attend or bring in documents.",
-  GovukHeadingM = "Get help paying court fees",
-  GovukBody2 = "You can check the help with fees guidance on ",
-  GovukLink1 = "GOV.UK (opens in a new tab)",
-  GovukBody3 = "to find out if you are eligible for support.",
-  next = "Start now",
-}
