--- e2e/fixtures/citizen/caseView/makeRequestToCourtAboutCase/applicant/requestToOrderWitnessContent4.ts
@@ -1,8 +1,0 @@
-export enum RequestToOrderWitnessContent4 {
-  GovukCaptionL = "Make a request to order a witness to attend court",
-  GovukHeadingL = "Do you have supporting documents to upload?",
-  GovukFieldsetLegend = "If you can show you have proof the other person in the case agrees to your request or you have any other supporting documents, you can upload them here.",
-  GovukLabel = "Yes",
-  GovukLabel1 = "No",
-  next = "Start now",
-}
