--- e2e/fixtures/citizen/caseView/makeRequestToCourtAboutCase/applicant/requestToOrderWitnessContent5.ts
@@ -1,16 +1,0 @@
-export enum RequestToOrderWitnessContent5 {
-  GovukCaptionL = "Make a request to order a witness to attend court",
-  GovukHeadingL = "Upload your supporting documents",
-  GovukBody = "If you are uploading a paper copy of the document, make sure this has been scanned in clearly and saved in a suitable format such as PDF.",
-  GovukHeadingS = "Upload documents",
-  GovukHint = "Give each document a file name that makes it clear what it is about. For example position-statement.docx. Files must end with JPG, JPEG, BMP, PNG, TIF, PDF, DOC or DOCX.",
-  GovukSummaryText = "Take a picture of a document on your phone and upload it",
-  GovukDetailsText1 = "Place your document on a flat service in a well-lit room. Use a flash if you need to.",
-  GovukDetailsText2 = "Take a picture of the whole document. You should be able to see its edges.",
-  GovukDetailsText3 = "Check you can read all the writing, including the handwriting.",
-  GovukDetailsText4 = "Email or send the photo or scan to the device you are using now.",
-  GovukDetailsText5 = "Upload it here.",
-  uploadButton = "Upload file",
-  removeButton = "Remove",
-  next = "Continue",
-}
