--- e2e/fixtures/citizen/caseView/makeRequestToCourtAboutCase/applicant/requestToOrderWitnessContent6.ts
@@ -1,10 +1,0 @@
-export enum RequestToOrderWitnessContent6 {
-  GovukCaptionL = "Make a request to order a witness to attend court",
-  GovukHeadingM = "Is there a reason why your request needs to be considered in the next five days?",
-  GovukLabel = "Yes",
-  GovukLabel1 = "No",
-  GovukLabel2 = "Give a reason why the court should consider your request urgently",
-  GovukHint = "For example, if there is an upcoming hearing or deadline set by the court.",
-  reasonForUrgencyInput = "Testing reason why this request is urgent",
-  next = "Continue",
-}
