--- e2e/fixtures/citizen/caseView/makeRequestToCourtAboutCase/applicant/requestToOrderWitnessContentCYA.ts
@@ -1,11 +1,0 @@
-export enum RequestToOrderWitnessContentCYA {
-  GovukHeadingXl = "Check your answers",
-  GovukSummaryListKey1 = "What are you applying for?",
-  GovukSummaryListKey2 = "Document uploaded",
-  GovukSummaryListKey3 = "Do you have supporting documents to upload?",
-  GovukSummaryListValue1 = "FP25 application",
-  GovukSummaryListValue2 = "mockFile.pdf",
-  GovukSummaryListValue3 = "Yes",
-  GovukLink = "Change",
-  submit = "Submit Application",
-}
