+++ e2e/fixtures/citizen/caseView/makeRequestToCourtAboutCase/listOfApplications1Content.ts
