+++ e2e/fixtures/citizen/caseView/makeRequestToCourtAboutCase/requestToOrderWitnessContent3.ts
@@ -1,0 +1,6 @@
+export enum RequestToOrderWitnessContent3 {
+  caption = "Make a request to order a witness to attend court",
+  heading = "Help with fees",
+  paragraph = "The cost of this application is £60.00. You can check the help with fees guidance on GOV.UK (opens in a new tab) to find out if you are eligible for support.",
+  radioLegend = "Will you be using help with fees to pay for this application?",
+}
