--- e2e/fixtures/citizen/caseView/makeRequestToCourtAboutCase/respondent/listOfApplications1Content.ts
@@ -1,23 +1,0 @@
-export enum ListOfApplications1Content {
-  GovukHeadingL = "Make a request to the court about your case",
-  GovukBody1 = "Select a form to make an application in your court proceedings.",
-  GovukAccordionSection1 = "Ask to delay or cancel a hearing date",
-  GovukBody2 = "You can apply to delay or cancel a hearing by completing and submitting the form C2.",
-  GovukLinkC2 = "Apply to the court using form C2",
-  GovukAccordionSection2 = "Request more time to do what is required by a court order",
-  GovukBody3 = "You can ask for more time to take the actions set out by a court order by completing and submitting the form C2.",
-  GovukAccordionSection3 = "Apply to change, extend or cancel a non-molestation order or occupation order",
-  GovukBody4 = "Changes, extensions or cancelling an order can be requested by completing and submitting the form FL403.",
-  GovukLink1 = "Apply to the court using form FL403",
-  GovukAccordionSection4 = "Other requests to the court where you need to complete a form C2",
-  GovukBody5 = "You can complete form C2 to request the following:",
-  GovukLink2 = "Ask to submit evidence the court has not requested",
-  GovukLink3 = "Ask to share documents with someone else",
-  GovukLink4 = "Ask to join or leave a case",
-  GovukLink5 = "Request to withdraw an application",
-  GovukLink6 = "Ask the court to appoint an expert (such as a medical professional or a child psychologist)",
-  GovukAccordionSection5 = "Appeal a court order or ask for permission to appeal",
-  GovukBody6 = "You can appeal or ask for permission to appeal a court order by completing and submitting and completing form N161.",
-  GovukLink7 = "Apply to the court using form N161",
-  next = "2 of 2",
-}
