--- e2e/fixtures/citizen/caseView/makeRequestToCourtAboutCase/respondent/requestToOrderWitnessContent2.ts
@@ -1,9 +1,0 @@
-export enum RequestToOrderWitnessContent2 {
-  GovukCaptionL = "Make a request to order a witness to attend court",
-  GovukHeadingL = "Upload your application",
-  GovukBodyM = "You will need to fill in the form FP25 and upload it when submitting this request.",
-  GovukFieldsetLegend = "Have you already completed the FP25 form?",
-  GovukLabel1 = "Yes",
-  GovukLabel2 = "No",
-  next = "Continue",
-}
