--- e2e/fixtures/citizen/caseView/makeRequestToCourtAboutCase/respondent/requestToOrderWitnessContent3.ts
@@ -1,11 +1,0 @@
-export enum RequestToOrderWitnessContent3 {
-  GovukCaptionL = "Make a request to order a witness to attend court",
-  GovukHeadingL = "Help with fees",
-  GovukBodyM = "The cost of this application is £60.00. You can check the help with fees guidance on",
-  GovukLink = "GOV.UK (opens in a new tab)",
-  GovukBodyM2 = "to find out if you are eligible for support.",
-  GovukFieldsetLegend = "Will you be using help with fees to pay for this application?",
-  GovukLabel = "Yes",
-  GovukLabel1 = "No",
-  continue = "Continue",
-}
