+++ e2e/pages/citizen/caseView/makeRequestToCourtAboutCase/listOfApplications1Page.ts
@@ -1,8 +1,8 @@
 import { Page } from "@playwright/test";
-import { Selectors } from "../../../../../common/selectors.ts";
+import { Selectors } from "../../../../common/selectors.ts";
 import { AxeUtils } from "@hmcts/playwright-common";
-import { ListOfApplications1Content } from "../../../../../fixtures/citizen/caseView/makeRequestToCourtAboutCase/applicant/listOfApplications1Content.ts";
-import { Helpers } from "../../../../../common/helpers.ts";
+import { ListOfApplications1Content } from "../../../../fixtures/citizen/caseView/makeRequestToCourtAboutCase/listOfApplications1Content.ts";
+import { Helpers } from "../../../../common/helpers.ts";
 
 export class ListOfApplications1Page {
   public static async listOfApplications1Page(
