+++ e2e/pages/citizen/caseView/makeRequestToCourtAboutCase/requestToOrderWitnessToAttendCourtPage5.ts
@@ -1,17 +1,17 @@
 import { Page } from "@playwright/test";
-import { Selectors } from "../../../../../common/selectors.ts";
-import { Helpers } from "../../../../../common/helpers.ts";
+import { Selectors } from "../../../../common/selectors.ts";
+import { Helpers } from "../../../../common/helpers.ts";
 import { AxeUtils } from "@hmcts/playwright-common";
-import { CommonStaticText } from "../../../../../common/commonStaticText.ts";
-import { RequestToOrderWitnessContent4 } from "../../../../../fixtures/citizen/caseView/makeRequestToCourtAboutCase/applicant/requestToOrderWitnessContent4.ts";
+import { CommonStaticText } from "../../../../common/commonStaticText.ts";
+import { RequestToOrderWitnessContent5 } from "../../../../fixtures/citizen/caseView/makeRequestToCourtAboutCase/requestToOrderWitnessContent5.ts";
 
 enum UniqueSelectors {
   supportingDocuments_Yes = "#awp_hasSupportingDocuments",
   supportingDocuments_No = "#awp_hasSupportingDocuments-2",
 }
 
-export class RequestToOrderWitnessToAttendCourtPage4 {
-  public static async requestToOrderWitnessToAttendCourtPage4(
+export class RequestToOrderWitnessToAttendCourtPage5 {
+  public static async requestToOrderWitnessToAttendCourtPage5(
     page: Page,
     accessibilityTest: boolean,
     haveSupportingDocuments: boolean,
@@ -26,29 +26,29 @@ export class RequestToOrderWitnessToAttendCourtPage4 {
   ): Promise<void> {
     await page
       .locator(Selectors.GovukCaptionL, {
-        hasText: RequestToOrderWitnessContent4.GovukCaptionL,
+        hasText: RequestToOrderWitnessContent5.GovukCaptionL,
       })
       .waitFor();
 
     await Promise.all([
       Helpers.checkVisibleAndPresent(
         page,
-        `${Selectors.GovukHeadingL}:text-is("${RequestToOrderWitnessContent4.GovukHeadingL}")`,
+        `${Selectors.GovukHeadingL}:text-is("${RequestToOrderWitnessContent5.GovukHeadingL}")`,
         1,
       ),
       Helpers.checkVisibleAndPresent(
         page,
-        `${Selectors.GovukFieldsetLegend}:text-is("${RequestToOrderWitnessContent4.GovukFieldsetLegend}")`,
+        `${Selectors.GovukFieldsetLegend}:text-is("${RequestToOrderWitnessContent5.GovukFieldsetLegend}")`,
         1,
       ),
       Helpers.checkVisibleAndPresent(
         page,
-        `${Selectors.GovukLabel}:text-is("${RequestToOrderWitnessContent4.GovukLabel}")`,
+        `${Selectors.GovukLabel}:text-is("${RequestToOrderWitnessContent5.GovukLabel}")`,
         1,
       ),
       Helpers.checkVisibleAndPresent(
         page,
-        `${Selectors.GovukLabel}:text-is("${RequestToOrderWitnessContent4.GovukLabel1}")`,
+        `${Selectors.GovukLabel}:text-is("${RequestToOrderWitnessContent5.GovukLabel1}")`,
         1,
       ),
     ]);
