+++ e2e/pages/citizen/caseView/makeRequestToCourtAboutCase/requestToOrderWitnessToAttendCourtPage6.ts
@@ -1,11 +1,11 @@
 import { Page } from "@playwright/test";
-import { Selectors } from "../../../../../common/selectors.ts";
-import { Helpers } from "../../../../../common/helpers.ts";
+import { Selectors } from "../../../../common/selectors.ts";
+import { Helpers } from "../../../../common/helpers.ts";
 import { AxeUtils } from "@hmcts/playwright-common";
-import { CommonStaticText } from "../../../../../common/commonStaticText.ts";
-import { RequestToOrderWitnessContent3 } from "../../../../../fixtures/citizen/caseView/makeRequestToCourtAboutCase/applicant/requestToOrderWitnessContent3.ts";
-import config from "../../../../../utils/config.utils.ts";
-import { RequestToOrderWitnessContent5 } from "../../../../../fixtures/citizen/caseView/makeRequestToCourtAboutCase/applicant/requestToOrderWitnessContent5.ts";
+import { CommonStaticText } from "../../../../common/commonStaticText.ts";
+import { RequestToOrderWitnessContent4 } from "../../../../fixtures/citizen/caseView/makeRequestToCourtAboutCase/requestToOrderWitnessContent4.ts";
+import config from "../../../../utils/config.utils.ts";
+import { RequestToOrderWitnessContent6 } from "../../../../fixtures/citizen/caseView/makeRequestToCourtAboutCase/requestToOrderWitnessContent6.ts";
 
 interface UploadSupportingDocuments {
   page: Page;
@@ -20,7 +20,7 @@ enum UniqueSelectors {
   documentUpload = "#awp-doc-form-upload",
 }
 
-export class RequestToOrderWitnessToAttendCourtPage5 {
+export class RequestToOrderWitnessToAttendCourtPage6 {
   public static async uploadSupportingDocumentsPage({
     page,
     accessibilityTest,
@@ -35,43 +35,43 @@ export class RequestToOrderWitnessToAttendCourtPage5 {
   ): Promise<void> {
     await page
       .locator(Selectors.GovukCaptionL, {
-        hasText: RequestToOrderWitnessContent5.GovukCaptionL,
+        hasText: RequestToOrderWitnessContent6.GovukCaptionL,
       })
       .waitFor();
     await Promise.all([
       Helpers.checkVisibleAndPresent(
         page,
-        `${Selectors.GovukHeadingL}:text-is("${RequestToOrderWitnessContent5.GovukHeadingL}")`,
+        `${Selectors.GovukHeadingL}:text-is("${RequestToOrderWitnessContent6.GovukHeadingL}")`,
         1,
       ),
       Helpers.checkVisibleAndPresent(
         page,
-        `${Selectors.GovukBody}:text-is("${RequestToOrderWitnessContent5.GovukBody}")`,
+        `${Selectors.GovukBody}:text-is("${RequestToOrderWitnessContent6.GovukBody}")`,
         1,
       ),
       Helpers.checkVisibleAndPresent(
         page,
-        `${Selectors.GovukHeadingS}:text-is("${RequestToOrderWitnessContent5.GovukHeadingS}")`,
+        `${Selectors.GovukHeadingS}:text-is("${RequestToOrderWitnessContent6.GovukHeadingS}")`,
         1,
       ),
       Helpers.checkVisibleAndPresent(
         page,
-        `${Selectors.GovukHint}:text-is("${RequestToOrderWitnessContent5.GovukHint}")`,
+        `${Selectors.GovukHint}:text-is("${RequestToOrderWitnessContent6.GovukHint}")`,
         1,
       ),
       Helpers.checkVisibleAndPresent(
         page,
-        `${Selectors.GovukSummaryText}:text-is("${RequestToOrderWitnessContent5.GovukSummaryText}")`,
+        `${Selectors.GovukSummaryText}:text-is("${RequestToOrderWitnessContent6.GovukSummaryText}")`,
         1,
       ),
       await page.click(
-        `${Selectors.GovukSummaryText}:text-is("${RequestToOrderWitnessContent5.GovukSummaryText}")`,
+        `${Selectors.GovukSummaryText}:text-is("${RequestToOrderWitnessContent6.GovukSummaryText}")`,
       ),
     ]);
     await Helpers.checkGroup(
       page,
       5,
-      RequestToOrderWitnessContent3,
+      RequestToOrderWitnessContent4,
       `GovukDetailsText`,
       `${Selectors.li}`,
     );
@@ -86,11 +86,11 @@ export class RequestToOrderWitnessToAttendCourtPage5 {
     await fileInput.setInputFiles(config.testPdfFile);
 
     await page.click(
-      `${Selectors.GovukButton}:text-is("${RequestToOrderWitnessContent5.uploadButton}")`,
+      `${Selectors.GovukButton}:text-is("${RequestToOrderWitnessContent6.uploadButton}")`,
     );
     await Helpers.checkVisibleAndPresent(
       page,
-      `${Selectors.a}:text-is("${RequestToOrderWitnessContent5.removeButton}")`,
+      `${Selectors.a}:text-is("${RequestToOrderWitnessContent6.removeButton}")`,
       1,
     );
     await page.click(
