+++ e2e/tests/citizen/caseView/makeRequestToCourtAboutCase/respondent/makeRequestToOrderWitnessAttendCourt.spec.ts
@@ -1,41 +1,41 @@
 import { test } from "../../../../fixtures.ts";
 import config from "../../../../../utils/config.utils.ts";
-import { Helpers } from "../../../../../common/helpers.ts";
-import { RequestToOrderWitnessToAttendCourt } from "../../../../../journeys/citizen/caseView/makeRequestToCourtAboutCase/respondent/requestToOrderWitnessToAttendCourt.ts";
+import { RequestToOrderWitnessToAttendCourt } from "../../../../../journeys/citizen/caseView/makeRequestToCourtAboutCase/requestToOrderWitnessToAttendCourt.ts";
+import { ActivateCitizenC100Case } from "../../../../../journeys/citizen/activateCase/activateCitizenC100Case.ts";
 
 test.use({ storageState: config.sessionStoragePath + "caseWorker.json" });
 
 test.describe("Respondent Make a request to order a witness to attend court tests", (): void => {
-  test.slow();
-  let ccdRef: string;
-
-  test.beforeEach(async ({ page, courtNavUtils }) => {
-    ccdRef = await courtNavUtils.createCase(true, false);
-    await Helpers.goToCase(
+  test.beforeEach(
+    async ({
       page,
-      config.manageCasesBaseURLCase,
-      ccdRef,
-      "tasks",
-    );
-  });
+      citizenC100CaseUtils,
+      idamLoginHelper,
+      accessCodeHelper,
+    }) => {
+      await ActivateCitizenC100Case.activateCase({
+        page,
+        citizenC100CaseUtils,
+        idamLoginHelper,
+        accessCodeHelper,
+        isApplicant: false,
+      });
+    },
+  );
 
   test(
     "Respondent Make a request to order a witness to attend court page." +
       "No Help With Fees." +
       "Already Completed FP25 @regression @accessibility @nightly",
-    async ({ page, browser }): Promise<void> => {
+    async ({ page }): Promise<void> => {
       await RequestToOrderWitnessToAttendCourt.requestToOrderWitnessToAttendCourt(
         {
           page: page,
-          browser: browser,
-          caseRef: ccdRef,
           accessibilityTest: false,
-          isApplicant: false,
           alreadyCompletedFP25: true,
           haveSupportingDocumentsUpload: true,
-          usingHwf: false,
+          helpWithFees: false,
           reasonForUrgency: true,
-          applicationSubmittedBy: "Citizen",
         },
       );
     },
